Split place and map utils by concern, drop dead opening-hours parser
