+++ web/src/lib/map/mapbox-style-compat.ts
@@ -1,9 +1,4 @@
 import { Layer, MaplibreLayerType } from '@/types/map.types'
-import dayjs from 'dayjs'
-import localizedFormat from 'dayjs/plugin/localizedFormat'
-
-// Initialize the plugin
-dayjs.extend(localizedFormat)
 
 // List of Mapbox paint properties that are not supported by Maplibre
 const MAPBOX_PAINT_PROPERTIES = [
@@ -180,220 +175,4 @@ export function mapboxLayerToMaplibreLayer(
     ...layer,
     configuration: maplibreConfig,
   } as any
-}
-
-export function parseMapboxToOsmId(featureId: string | number): {
-  osmId: string
-  poiType: 'node' | 'way' | 'relation' | 'unknown'
-} {
-  const typeCode = String(featureId).slice(-1)
-  const osmId = String(featureId).slice(0, -1)
-
-  const poiTypeCodeMap: {
-    [key: string]: 'node' | 'way' | 'relation' | 'unknown'
-  } = {
-    '0': 'node',
-    '1': 'way',
-    '2': 'unknown',
-    '3': 'unknown',
-    '4': 'relation',
-  }
-
-  return {
-    osmId,
-    poiType: poiTypeCodeMap[typeCode] || 'unknown',
-  }
-}
-
-/**
- * Parse a Planetiler/OpenMapTiles MVT feature ID into an OSM ID and type.
- * Planetiler encodes as: feature.id = osm_id * 10 + type_code
- * where type_code: 1=node, 2=way, 3=relation
- */
-export function parsePlanetilerOsmId(featureId: string | number): {
-  osmId: string
-  poiType: 'node' | 'way' | 'relation' | 'unknown'
-} {
-  const id = typeof featureId === 'string' ? parseInt(featureId, 10) : featureId
-  if (isNaN(id) || id <= 0) return { osmId: '0', poiType: 'unknown' }
-
-  const typeCode = id % 10
-  const osmId = Math.floor(id / 10)
-
-  const poiTypeCodeMap: Record<number, 'node' | 'way' | 'relation'> = {
-    1: 'node',
-    2: 'way',
-    3: 'relation',
-  }
-
-  return {
-    osmId: String(osmId),
-    poiType: poiTypeCodeMap[typeCode] || 'unknown',
-  }
-}
-
-type OpeningStatus = {
-  isOpen: boolean
-  nextChange: string | null
-}
-
-type DayRange = {
-  days: string[]
-  hours: string[]
-}
-
-export function parseOpeningHours(hoursStr: string, timezone?: string): OpeningStatus | null {
-  try {
-    // Handle special cases first
-    if (hoursStr === '24/7') {
-      return {
-        isOpen: true,
-        nextChange: null, // Always open
-      }
-    }
-
-    if (hoursStr.toLowerCase() === 'sunrise-sunset') {
-      return {
-        isOpen: true,
-        nextChange: 'at sunset',
-      }
-    }
-
-    let now: dayjs.Dayjs
-    let currentDay: string
-    let currentTime: string
-    let currentDayNum: number
-
-    if (timezone) {
-      const date = new Date()
-      const parts = new Intl.DateTimeFormat('en-US', {
-        timeZone: timezone,
-        hour: '2-digit',
-        minute: '2-digit',
-        hour12: false,
-        weekday: 'short',
-      }).formatToParts(date)
-      const weekdayStr = parts.find(p => p.type === 'weekday')?.value ?? ''
-      const dayMap: Record<string, number> = { Sun: 0, Mon: 1, Tue: 2, Wed: 3, Thu: 4, Fri: 5, Sat: 6 }
-      currentDayNum = dayMap[weekdayStr] ?? date.getDay()
-      currentDay = dayjs().day(currentDayNum).format('dd').slice(0, 2)
-      const hour = parts.find(p => p.type === 'hour')?.value ?? '00'
-      const minute = parts.find(p => p.type === 'minute')?.value ?? '00'
-      currentTime = `${hour.padStart(2, '0')}:${minute.padStart(2, '0')}`
-    } else {
-      now = dayjs()
-      currentDay = now.format('dd').slice(0, 2)
-      currentTime = now.format('HH:mm')
-      currentDayNum = now.day()
-    }
-
-    // Split into day ranges
-    const ranges = hoursStr.split(';').map(range => {
-      // Handle case where there's no day specified (just hours)
-      if (!range.includes(' ')) {
-        return {
-          days: ['Mo-Su'],
-          hours: range.trim().split('-'),
-        }
-      }
-
-      const [days, hours] = range.trim().split(' ')
-      // If there's only one range and no day specified, assume it's every day
-      const parsedDays =
-        days.includes('-') || days.includes(',') ? days.split(',') : ['Mo-Su']
-      return {
-        days: parsedDays,
-        hours: hours === 'off' ? null : hours.split('-'),
-      }
-    })
-
-    // Find current day's hours
-    const todayRange = ranges.find(range =>
-      range.days.some(day => {
-        // Handle the Mo-Su case
-        if (day === 'Mo-Su') return true
-        return day.includes(currentDay)
-      }),
-    )
-
-    // Helper to format time
-    const formatTime = (time: string) => {
-      const [hours, minutes] = time.split(':')
-      return dayjs()
-        .hour(parseInt(hours))
-        .minute(parseInt(minutes))
-        .format('h:mm A')
-    }
-
-    // Helper to find next open day and its hours
-    const findNextOpenDay = (
-      startDay: number,
-    ): { day: number; hours: string[] } | null => {
-      for (let i = 1; i <= 7; i++) {
-        const nextDay = (startDay + i) % 7
-        const nextDayCode = dayjs().day(nextDay).format('dd').slice(0, 2)
-        const nextDayRange = ranges.find(range =>
-          range.days.some(day => day.includes(nextDayCode)),
-        )
-        if (nextDayRange?.hours) {
-          return { day: nextDay, hours: nextDayRange.hours }
-        }
-      }
-      return null
-    }
-
-    if (!todayRange || !todayRange.hours) {
-      // Place is closed today, find next opening
-      const nextOpen = findNextOpenDay(currentDayNum)
-      if (nextOpen) {
-        const nextOpenDate = dayjs()
-          .day(nextOpen.day)
-          .hour(parseInt(nextOpen.hours[0].split(':')[0]))
-          .minute(parseInt(nextOpen.hours[0].split(':')[1]))
-        return {
-          isOpen: false,
-          nextChange: nextOpenDate.format('dddd [at] h:mm A'),
-        }
-      }
-      return {
-        isOpen: false,
-        nextChange: null,
-      }
-    }
-
-    const [openTime, closeTime] = todayRange.hours
-    const isOpen = currentTime >= openTime && currentTime <= closeTime
-
-    if (isOpen) {
-      return {
-        isOpen: true,
-        nextChange: formatTime(closeTime),
-      }
-    } else if (currentTime < openTime) {
-      return {
-        isOpen: false,
-        nextChange: formatTime(openTime),
-      }
-    } else {
-      // Find next day's opening time
-      const nextOpen = findNextOpenDay(currentDayNum)
-      if (nextOpen) {
-        const nextOpenDate = dayjs()
-          .day(nextOpen.day)
-          .hour(parseInt(nextOpen.hours[0].split(':')[0]))
-          .minute(parseInt(nextOpen.hours[0].split(':')[1]))
-        return {
-          isOpen: false,
-          nextChange: nextOpenDate.format('dddd [at] h:mm A'),
-        }
-      }
-      return {
-        isOpen: false,
-        nextChange: null,
-      }
-    }
-  } catch (error) {
-    console.error('Failed to parse opening hours:', error)
-    return null
-  }
 }
