+++ web/src/lib/place/place-details.ts
@@ -1,0 +1,140 @@
+import { Place } from '@/types/place.types'
+
+export function formatAddress(place: Place): string {
+  if (!place.address) return ''
+
+  const address = place.address.value
+  const { street1, street2, locality, postalCode, neighborhood, region } =
+    address
+
+  const parts: string[] = []
+
+  // Add street address if available
+  const streetParts = [street1, street2].filter(Boolean)
+  const hasStreetAddress = streetParts.length > 0
+  if (hasStreetAddress) {
+    parts.push(streetParts.join(' '))
+  }
+
+  // Add neighborhood and locality
+  if (neighborhood && locality) {
+    parts.push(`${neighborhood}, ${locality}`)
+  } else if (neighborhood) {
+    parts.push(neighborhood)
+    if (locality) parts.push(locality)
+  } else if (locality) {
+    parts.push(locality)
+  }
+
+  // Add postal code if we don't have a street address
+  if (postalCode && !hasStreetAddress) {
+    parts.push(postalCode)
+  }
+  // Or add region if we don't have street address or postal code
+  else if (region && !hasStreetAddress && !postalCode) {
+    parts.push(region)
+  }
+
+  return parts.join(', ')
+}
+
+export function parseCuisines(cuisine: string | undefined): string[] | null {
+  if (!cuisine) return null
+
+  return cuisine
+    .split(';')
+    .map(c => c.trim())
+    .map(c => c.replace(/_/g, ' '))
+    .map(c => c.charAt(0).toUpperCase() + c.slice(1))
+}
+
+export function getWifiStatus(tags: Record<string, string | undefined>) {
+  const access = tags.internet_access
+  const ssid = tags['internet_access:ssid']
+  const fee = tags['internet_access:fee']
+  const password = tags['internet_access:password']
+
+  if (!access || access === 'no') return null
+
+  let label = 'WiFi available'
+
+  if (access === 'free' || fee === 'no') {
+    label = 'Free WiFi available'
+  } else if (access === 'customers') {
+    label = 'WiFi for customers'
+  } else if (fee === 'yes') {
+    label = 'Paid WiFi available'
+  }
+
+  return {
+    label,
+    ssid,
+    password,
+  }
+}
+
+export function hasOutdoorSeating(
+  tags: Record<string, string | undefined>,
+): boolean {
+  return tags.outdoor_seating === 'yes'
+}
+
+export function getWheelchairAccess(
+  tags: Record<string, string | undefined>,
+): string {
+  const wheelchair = tags?.wheelchair || 'unknown'
+
+  switch (wheelchair) {
+    case 'yes':
+      return 'Wheelchair accessible'
+    case 'no':
+      return 'Not wheelchair accessible'
+    case 'limited':
+      return 'Limited wheelchair accessibility'
+    case 'designated':
+      return 'Designated wheelchair access'
+    default:
+      return 'Unknown wheelchair accessibility'
+  }
+}
+
+export function getSmokingStatus(
+  tags: Record<string, string | undefined>,
+): string {
+  const smoking = tags?.smoking || 'unknown'
+
+  switch (smoking) {
+    case 'yes':
+      return 'Smoking allowed'
+    case 'no':
+      return 'No smoking'
+    case 'separated':
+      return 'Separate smoking area'
+    case 'isolated':
+      return 'Isolated smoking area'
+    case 'outside':
+      return 'Smoking allowed outside'
+    case 'dedicated':
+      return 'Dedicated smoking area'
+    default:
+      return 'Unknown smoking policy'
+  }
+}
+
+export function getRestroomAccess(
+  tags: Record<string, string | undefined>,
+): string {
+  const toilets = tags?.toilets || 'unknown'
+
+  switch (toilets) {
+    case 'yes':
+      return 'Restrooms available'
+    case 'no':
+      return 'No restrooms'
+    case 'customers':
+      return 'Restrooms for customers only'
+    default:
+      return 'Unknown restroom availability'
+  }
+}
+
