+++ web/src/services/map/providers/mapbox.strategy.ts
@@ -37,7 +37,7 @@ import { decodeShape } from '@/lib/utils'
 import { palette } from '@/lib/palette'
 import { mapEventBus } from '@/lib/event-bus'
 import { createPegmanLayers, updatePegmanData } from '@/lib/street-view/pegman'
-import { parseMapboxToOsmId } from '@/lib/map/map.utils'
+import { parseMapboxToOsmId } from '@/lib/map/osm-feature-id'
 import { useRouter } from 'vue-router'
 import { AppRoute } from '@/router'
 import { MapLayerGroup, TripGroup } from '@/services/map/providers/layer-group'
