+++ web/src/services/place/place-display.ts
@@ -6,7 +6,8 @@ import type { RecentPlaceEntry, RecentSearchEntry } from '@/lib/recents'
 import type { AutocompleteResult } from '@/types/search.types'
 import type { ThemeColor } from '@/lib/utils'
 import { AppRoute } from '@/router'
-import { getPlaceRoute, getPlaceRouteFromExternalIds, getTransitStopRoute, formatAddress } from '@/lib/place/place.utils'
+import { getPlaceRoute, getPlaceRouteFromExternalIds, getTransitStopRoute } from '@/lib/place/place-route'
+import { formatAddress } from '@/lib/place/place-details'
 import { getCategoryColor } from '@/services/place/place-colors'
 import { frequentChipMeta } from '@/lib/frequents'
 import {
