+++ web/src/services/search-suggestions.service.ts
@@ -11,7 +11,7 @@ import type { PlaceCategory } from '@/types/place.types'
 import { COMMON_CATEGORIES } from '@/lib/place/common-categories'
 import { CommandArgumentOption } from '@/types/command.types'
 import { frequentChipMeta } from '@/lib/frequents'
-import { getBookmarkPlaceId } from '@/lib/place/place.utils'
+import { getBookmarkPlaceId } from '@/lib/place/place-route'
 import { getCategoryColor } from '@/services/place/place-colors'
 import { recentPlaceIdentity, recentSearchIdentity, RecentSearchEntry, RecentPlaceEntry } from '@/lib/recents'
 import { useBookmarksStore } from '@/stores/library/bookmarks.store'
