+++ web/src/views/search/Search.vue
@@ -16,7 +16,7 @@ import PlaceList from '@/components/place/card/PlaceList.vue'
 import FilterChips from '@/components/map/FilterChips.vue'
 import ErrorMessage from '@/components/ErrorMessage.vue'
 import { useRouter } from 'vue-router'
-import { getPlaceRoute } from '@/lib/place/place.utils'
+import { getPlaceRoute } from '@/lib/place/place-route'
 import { usePlaceService } from '@/services/place/place.service'
 import { useRecentsStore } from '@/stores/recents.store'
 import { useCategoryStore } from '@/stores/category.store'
